move cli to main
